♻️ forward declare some important structs in the `Definitions` header and move the actual definitions where they belong ♻️ introduce a separate header for `Edge` and `Control` to better separate the code into smaller chunks ✨ std::hash overloads for `Complex`, `ComplexValue` and `Edge` ♻️ compute table no longer requires a hash function as template argument, but uses std::hash
