+++ src/CMakeLists.txt
@@ -3,7 +3,9 @@ add_library(${PROJECT_NAME}
             ${CMAKE_CURRENT_SOURCE_DIR}/DDpackage.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/DDcomplex.cpp
             ${CMAKE_CURRENT_SOURCE_DIR}/DDexport.cpp
+            ${${PROJECT_NAME}_SOURCE_DIR}/include/Control.hpp
             ${${PROJECT_NAME}_SOURCE_DIR}/include/Definitions.hpp
+            ${${PROJECT_NAME}_SOURCE_DIR}/include/Edge.hpp
             ${${PROJECT_NAME}_SOURCE_DIR}/include/GateMatrixDefinitions.hpp
             ${${PROJECT_NAME}_SOURCE_DIR}/include/UniqueTable.hpp
             ${${PROJECT_NAME}_SOURCE_DIR}/include/ComputeTable.hpp
